Trait massa_api::Endpoints

pub trait Endpoints: Sized + Send + Sync + 'static {

    fn stop_node(&self) -> BoxFuture<Result<(), ApiError>>;
    fn node_sign_message(
        &self,
        _: Vec<u8>
    ) -> BoxFuture<Result<PubkeySig, ApiError>>;
    fn add_staking_secret_keys(
        &self,
        _: Vec<String>
    ) -> BoxFuture<Result<(), ApiError>>;
    fn execute_read_only_bytecode(
        &self,
        _: Vec<ReadOnlyBytecodeExecution>
    ) -> BoxFuture<Result<Vec<ExecuteReadOnlyResponse>, ApiError>>;
    fn execute_read_only_call(
        &self,
        _: Vec<ReadOnlyCall>
    ) -> BoxFuture<Result<Vec<ExecuteReadOnlyResponse>, ApiError>>;
    fn remove_staking_addresses(
        &self,
        _: Vec<Address>
    ) -> BoxFuture<Result<(), ApiError>>;
    fn get_staking_addresses(
        &self
    ) -> BoxFuture<Result<PreHashSet<Address>, ApiError>>;
    fn node_ban_by_ip(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>;
    fn node_ban_by_id(&self, _: Vec<NodeId>) -> BoxFuture<Result<(), ApiError>>;
    fn node_whitelist(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>;
    fn node_remove_from_whitelist(
        &self,
        _: Vec<IpAddr>
    ) -> BoxFuture<Result<(), ApiError>>;
    fn node_unban_by_ip(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>;
    fn node_unban_by_id(&self, _: Vec<NodeId>) -> BoxFuture<Result<(), ApiError>>;
    fn get_status(&self) -> BoxFuture<Result<NodeStatus, ApiError>>;
    fn get_cliques(&self) -> BoxFuture<Result<Vec<Clique>, ApiError>>;
    fn get_stakers(&self) -> BoxFuture<Result<Vec<(Address, u64)>, ApiError>>;
    fn get_operations(
        &self,
        _: Vec<OperationId>
    ) -> BoxFuture<Result<Vec<OperationInfo>, ApiError>>;
    fn get_endorsements(
        &self,
        _: Vec<EndorsementId>
    ) -> BoxFuture<Result<Vec<EndorsementInfo>, ApiError>>;
    fn get_block(&self, _: BlockId) -> BoxFuture<Result<BlockInfo, ApiError>>;
    fn get_blockclique_block_by_slot(
        &self,
        _: Slot
    ) -> BoxFuture<Result<Option<Block>, ApiError>>;
    fn get_graph_interval(
        &self,
        _: TimeInterval
    ) -> BoxFuture<Result<Vec<BlockSummary>, ApiError>>;
    fn get_datastore_entries(
        &self,
        _: Vec<DatastoreEntryInput>
    ) -> BoxFuture<Result<Vec<DatastoreEntryOutput>, ApiError>>;
    fn get_addresses(
        &self,
        _: Vec<Address>
    ) -> BoxFuture<Result<Vec<AddressInfo>, ApiError>>;
    fn send_operations(
        &self,
        _: Vec<OperationInput>
    ) -> BoxFuture<Result<Vec<OperationId>, ApiError>>;
    fn get_filtered_sc_output_event(
        &self,
        _: EventFilter
    ) -> BoxFuture<Result<Vec<SCOutputEvent>, ApiError>>;
    fn get_openrpc_spec(&self) -> BoxFuture<Result<Value, ApiError>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }

}

Exposed API endpoints

Required Methods

fn stop_node(&self) -> BoxFuture<Result<(), ApiError>>

Gracefully stop the node.

fn node_sign_message(&self, _: Vec<u8>) -> BoxFuture<Result<PubkeySig, ApiError>>

Sign message with node’s key. Returns the public key that signed the message and the signature.

fn add_staking_secret_keys(
&self,
_: Vec<String>
) -> BoxFuture<Result<(), ApiError>>

Add a vector of new secret(private) keys for the node to use to stake. No confirmation to expect.

fn execute_read_only_bytecode(
&self,
_: Vec<ReadOnlyBytecodeExecution>
) -> BoxFuture<Result<Vec<ExecuteReadOnlyResponse>, ApiError>>

Execute bytecode in read-only mode.

fn execute_read_only_call(
&self,
_: Vec<ReadOnlyCall>
) -> BoxFuture<Result<Vec<ExecuteReadOnlyResponse>, ApiError>>

Execute an SC function in read-only mode.

fn remove_staking_addresses(
&self,
_: Vec<Address>
) -> BoxFuture<Result<(), ApiError>>

Remove a vector of addresses used to stake. No confirmation to expect.

fn get_staking_addresses(
&self
) -> BoxFuture<Result<PreHashSet<Address>, ApiError>>

Return hash set of staking addresses.

fn node_ban_by_ip(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>

Bans given IP address(es). No confirmation to expect.

fn node_ban_by_id(&self, _: Vec<NodeId>) -> BoxFuture<Result<(), ApiError>>

Bans given node id. No confirmation to expect.

fn node_whitelist(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>

whitelist given IP address. No confirmation to expect. Note: If the ip was unknown it adds it to the known peers, otherwise it updates the peer type

fn node_remove_from_whitelist(
&self,
_: Vec<IpAddr>
) -> BoxFuture<Result<(), ApiError>>

remove from whitelist given IP address. keep it as standard No confirmation to expect.

fn node_unban_by_ip(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>

Unban given IP address(es). No confirmation to expect.

fn node_unban_by_id(&self, _: Vec<NodeId>) -> BoxFuture<Result<(), ApiError>>

Unban given node id. No confirmation to expect.

fn get_status(&self) -> BoxFuture<Result<NodeStatus, ApiError>>

Summary of the current state: time, last final blocks (hash, thread, slot, timestamp), clique count, connected nodes count.

fn get_cliques(&self) -> BoxFuture<Result<Vec<Clique>, ApiError>>

Get cliques.

fn get_stakers(&self) -> BoxFuture<Result<Vec<(Address, u64)>, ApiError>>

Returns the active stakers and their active roll counts for the current cycle.

fn get_operations(
&self,
_: Vec<OperationId>
) -> BoxFuture<Result<Vec<OperationInfo>, ApiError>>

Returns operations information associated to a given list of operations’ IDs.

fn get_endorsements(
&self,
_: Vec<EndorsementId>
) -> BoxFuture<Result<Vec<EndorsementInfo>, ApiError>>

Get endorsements (not yet implemented).

fn get_block(&self, _: BlockId) -> BoxFuture<Result<BlockInfo, ApiError>>

Get information on a block given its hash.

fn get_blockclique_block_by_slot(
&self,
_: Slot
) -> BoxFuture<Result<Option<Block>, ApiError>>

Get information on the block at a slot in the blockclique. If there is no block at this slot a None is returned.

fn get_graph_interval(
&self,
_: TimeInterval
) -> BoxFuture<Result<Vec<BlockSummary>, ApiError>>

Get the block graph within the specified time interval. Optional parameters: from <time_start> (included) and to <time_end> (excluded) millisecond timestamp

fn get_datastore_entries(
&self,
_: Vec<DatastoreEntryInput>
) -> BoxFuture<Result<Vec<DatastoreEntryOutput>, ApiError>>

Get multiple datastore entries.

fn get_addresses(
&self,
_: Vec<Address>
) -> BoxFuture<Result<Vec<AddressInfo>, ApiError>>

Get addresses.

fn send_operations(
&self,
_: Vec<OperationInput>
) -> BoxFuture<Result<Vec<OperationId>, ApiError>>

Adds operations to pool. Returns operations that were ok and sent to pool.

fn get_filtered_sc_output_event(
&self,
_: EventFilter
) -> BoxFuture<Result<Vec<SCOutputEvent>, ApiError>>

Get events optionally filtered by:

start slot

end slot

emitter address

original caller address

operation id

fn get_openrpc_spec(&self) -> BoxFuture<Result<Value, ApiError>>

Get OpenRPC specification.

Provided Methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.