Trait massa_api::Endpoints

pub trait Endpoints: Sized + Send + Sync + 'static {

    fn stop_node(&self) -> BoxFuture<Result<(), ApiError>>;

    fn node_sign_message(
        &self, 
        _: Vec<u8>
    ) -> BoxFuture<Result<PubkeySig, ApiError>>;

    fn add_staking_private_keys(
        &self, 
        _: Vec<PrivateKey>
    ) -> BoxFuture<Result<(), ApiError>>;

    fn execute_read_only_request(
        &self, 
        _: Vec<ReadOnlyExecution>
    ) -> BoxFuture<Result<Vec<ExecuteReadOnlyResponse>, ApiError>>;

    fn remove_staking_addresses(
        &self, 
        _: Vec<Address>
    ) -> BoxFuture<Result<(), ApiError>>;

    fn get_staking_addresses(&self) -> BoxFuture<Result<Set<Address>, ApiError>>;

    fn ban(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>;

    fn unban(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>;

    fn get_status(&self) -> BoxFuture<Result<NodeStatus, ApiError>>;

    fn get_cliques(&self) -> BoxFuture<Result<Vec<Clique>, ApiError>>;

    fn get_stakers(&self) -> BoxFuture<Result<Map<Address, u64>, ApiError>>;

    fn get_operations(
        &self, 
        _: Vec<OperationId>
    ) -> BoxFuture<Result<Vec<OperationInfo>, ApiError>>;

    fn get_endorsements(
        &self, 
        _: Vec<EndorsementId>
    ) -> BoxFuture<Result<Vec<EndorsementInfo>, ApiError>>;

    fn get_block(&self, _: BlockId) -> BoxFuture<Result<BlockInfo, ApiError>>;

    fn get_graph_interval(
        &self, 
        _: TimeInterval
    ) -> BoxFuture<Result<Vec<BlockSummary>, ApiError>>;

    fn get_addresses(
        &self, 
        _: Vec<Address>
    ) -> BoxFuture<Result<Vec<AddressInfo>, ApiError>>;

    fn send_operations(
        &self, 
        _: Vec<SignedOperation>
    ) -> BoxFuture<Result<Vec<OperationId>, ApiError>>;

    fn get_filtered_sc_output_event(
        &self, 
        _: EventFilter
    ) -> BoxFuture<Result<Vec<SCOutputEvent>, ApiError>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }

}

Required methods

fn stop_node(&self) -> BoxFuture<Result<(), ApiError>>

Gracefully stop the node.

fn node_sign_message(
&self,
_: Vec<u8>
) -> BoxFuture<Result<PubkeySig, ApiError>>

Sign message with node’s key. Returns the public key that signed the message and the signature.

fn add_staking_private_keys(
&self,
_: Vec<PrivateKey>
) -> BoxFuture<Result<(), ApiError>>

Add a vec of new private keys for the node to use to stake. No confirmation to expect.

fn execute_read_only_request(
&self,
_: Vec<ReadOnlyExecution>
) -> BoxFuture<Result<Vec<ExecuteReadOnlyResponse>, ApiError>>

Execute code in read-only mode.

fn remove_staking_addresses(
&self,
_: Vec<Address>
) -> BoxFuture<Result<(), ApiError>>

Remove a vec of addresses used to stake. No confirmation to expect.

fn get_staking_addresses(&self) -> BoxFuture<Result<Set<Address>, ApiError>>

Return hashset of staking addresses.

fn ban(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>

Bans given IP address. No confirmation to expect.

fn unban(&self, _: Vec<IpAddr>) -> BoxFuture<Result<(), ApiError>>

Unbans given IP address. No confirmation to expect.

fn get_status(&self) -> BoxFuture<Result<NodeStatus, ApiError>>

Summary of the current state: time, last final blocks (hash, thread, slot, timestamp), clique count, connected nodes count.

fn get_cliques(&self) -> BoxFuture<Result<Vec<Clique>, ApiError>>

Get cliques.

fn get_stakers(&self) -> BoxFuture<Result<Map<Address, u64>, ApiError>>

Returns the active stakers and their active roll counts for the current cycle.

fn get_operations(
&self,
_: Vec<OperationId>
) -> BoxFuture<Result<Vec<OperationInfo>, ApiError>>

Returns operations information associated to a given list of operations’ IDs.

fn get_endorsements(
&self,
_: Vec<EndorsementId>
) -> BoxFuture<Result<Vec<EndorsementInfo>, ApiError>>

Get endorsements (not yet implemented).

fn get_block(&self, _: BlockId) -> BoxFuture<Result<BlockInfo, ApiError>>

Get information on a block given its hash.

fn get_graph_interval(
&self,
_: TimeInterval
) -> BoxFuture<Result<Vec<BlockSummary>, ApiError>>

Get the block graph within the specified time interval. Optional parameters: from <time_start> (included) and to <time_end> (excluded) millisecond timestamp

fn get_addresses(
&self,
_: Vec<Address>
) -> BoxFuture<Result<Vec<AddressInfo>, ApiError>>

Get addresses.

fn send_operations(
&self,
_: Vec<SignedOperation>
) -> BoxFuture<Result<Vec<OperationId>, ApiError>>

Adds operations to pool. Returns operations that were ok and sent to pool.

fn get_filtered_sc_output_event(
&self,
_: EventFilter
) -> BoxFuture<Result<Vec<SCOutputEvent>, ApiError>>

Get events optionnally filtered by:

start slot

end slot

emitter address

original caller address

operation id

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.